Split JS + TS arg type stories
